Refactor: TeamType을 global로 이동
