+++ src/main/java/ceos/study/vote/domain/leader/dto/LeaderRequestDTO.java
@@ -1,12 +1,12 @@
 package ceos.study.vote.domain.leader.dto;
 
-import ceos.study.vote.global.common.Part;
+import ceos.study.vote.global.common.PartType;
 import lombok.Getter;
 
 public class LeaderRequestDTO {
     @Getter
     public static class Vote{
         private Long id;
-        private Part part;
+        private PartType part;
     }
 }
