+++ src/main/java/ceos/study/vote/domain/leader/repository/LeaderRepository.java
@@ -1,11 +1,11 @@
 package ceos.study.vote.domain.leader.repository;
 
 import ceos.study.vote.domain.leader.entity.Leader;
-import ceos.study.vote.global.common.Part;
+import ceos.study.vote.global.common.PartType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface LeaderRepository extends JpaRepository<Leader, Long> {
-    List<Leader> findAllByPart(Part part);
+    List<Leader> findAllByPart(PartType part);
 }
