+++ src/main/java/ceos/study/vote/domain/team/repository/TeamRepository.java
@@ -1,7 +1,7 @@
 package ceos.study.vote.domain.team.repository;
 
 import ceos.study.vote.domain.team.entity.Team;
-import ceos.study.vote.domain.team.entity.TeamType;
+import ceos.study.vote.global.common.TeamType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
