+++ src/main/java/ceos/study/vote/domain/user/dto/UserRequestDto.java
@@ -1,5 +1,6 @@
 package ceos.study.vote.domain.user.dto;
 
+import ceos.study.vote.global.common.TeamType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
@@ -27,7 +28,7 @@ public class UserRequestDto {
 
         @NotEmpty
         @Schema(description = "팀 [INFLUY, HANIHOME, PROMESA, LOOPZ, DEARDREAM]", example = "INFLUY")
-        private String team;
+        private TeamType team;
 
         @NotEmpty
         @Schema(description = "파트", example = "백엔드")
