+++ src/main/java/ceos/study/vote/global/common/PartType.java
@@ -1,5 +1,5 @@
 package ceos.study.vote.global.common;
 
-public enum Part {
+public enum PartType {
     FE, BE;
 }
