--- src/main/java/ceos/study/vote/global/common/TeamName.java
@@ -1,5 +1,0 @@
-package ceos.study.vote.global.common;
-
-public enum TeamName {
-    INFLUY, HANIHOME, DREAM, LOOPZ, PROMESA;
-}
