+++ src/main/java/ceos/study/vote/global/common/TeamType.java
@@ -1,4 +1,4 @@
-package ceos.study.vote.domain.team.entity;
+package ceos.study.vote.global.common;
 
 public enum TeamType {
     INFLUY, HANIHOME, PROMESA, LOOPZ, DEARDREAM;
